refactor: extract bp_nemoclaw to routes/nemoclaw.py (final blueprint)
Phase 5.13 (FINAL) of the incremental modularisation. Moves the 7 NemoClaw
governance/approval routes into routes/nemoclaw.py. Module-level helpers
(_detect_nemoclaw, _parse_network_policies) and module state
(_nemoclaw_policy_hash, _nemoclaw_drift_info) stay in dashboard.py and
are accessed from handlers via late `import dashboard as _d`.

After this PR, zero `bp_XXX = _Blueprint(...)` declarations remain in
dashboard.py — all 22 Blueprints live in routes/.

Routes moved:
  GET  /api/nemoclaw/governance
  POST /api/nemoclaw/governance/acknowledge-drift
  GET  /api/nemoclaw/status
  GET  /api/nemoclaw/policy
  POST /api/nemoclaw/approve
  POST /api/nemoclaw/reject
  GET  /api/nemoclaw/pending-approvals
